refactor(musical): split MusicalTab.tsx into 4 sub-components (v3.3.2)
- MusicalParamsPanel: VibeBoard, tempo/metronome, instruments, rhythm, narrative
- MusicalPromptBuilder: generate button + prompt output panel
- LyricsMusicAnalysis: header + context band + guide steps
- MusicalTab: pure layout shell ~3KB
